Fix embedded requests when endpoint has a non-empty path
Version 2.3.1 introduced a problem with Artifactory endpoints where the endpoint
URL has a path part that is not empty, e.g., http://artifacts.example.com/artifactory.
In Resource::Base.from_url, we need to get the path relative to the endpoint, not
the entire path from the URL.

This fixes issue #60.