Add url attribute to support remote repositories
Added a url attribute to the repository resource, because it
allows reading and writing remote repositories.

Previously, to read the URL of a remote repository, we had to
make a raw GET request as in: /api/repositories/jcenter,
which would return the URL.

And previously, to create a remote repository, we had to make a
raw PUT request as in:
Artifactory.put('/api/repositories/myremote', json, headers)

With this new attribute, we can work with the Repository resource
directly.