Ensure read_timeout returns a non-zero value.
We inadvertently introduced a regression when converting the
ARTIFACTORY_READ_TIMEOUT value to a Fixnum. In the case where the
environment variable was not set, it would return `nil` which is
mutated to a `0` value when `nil.to_i` is run. This would cause all
connections to fail which did not have the value set as Net:HTTP would
not wait for a response before raising `Net::ReadTimeout`.